Differential Expression (DE) Analysis of sRNA dicer-derived clusters using `DESeq2` or `edgeR`

This function allows you to compute the differential expression of sRNA dicer-derived clusters. The function allows the choice between analysis with `DESeq2` or `edgeR`.

Usage

RNAanalysis(data, group, method = c("edgeR", "DESeq2"), dispersionValue = NULL)

Arguments

data

numeric data frame produced by [mobileRNA::RNAimport()] and/or [mobileRNA::RNAsubset()].

group

Vector of the condition (ie. treatment or control) for each sample. Must be stated in the same order as the samples in the `data` file from left to right.

method

The method to undertaken differential analysis, choose from methods of either [DESeq2::DESeq2] or [edgeR::edgeR]. Must be stated as either "DESeq2" or "edgeR" in the function.

dispersionValue

numeric value; manual setting of dispersion value which is recommended for analysis in experiments without biological replicates when utilising the [edgeR::edgeR] method.

Value

Undertakes differential analysis, based on a specified method, and adds the results to the supplied dataframe. This includes: * Count mean * Log fold change * p-value * Adjusted p-value * Log counts per million (CPM/RPM)

Details

The analysis allows the users to choose the method which best suits their data. Notably, `DESeq2` cannot compute the analysis when there only one replicate per condition, but, `edgeR` can. Simply set a suitable dispersion value, based on similar data, to use this feature. The dispersion value is other wise known as the common Biological squared coefficient of variation. A typical dispersion value is 0.4 for human data sets, 0.1 for data on genetically identical model organisms or 0.01 for technical replicate. See the User’s Guide for the ‘EdgeR’ package for more details, [edgeR::edgeR].

Examples

# sample conditions.
groups <- c("Tomato/Eggplant", "Tomato/Eggplant", "Tomato/Eggplant",
          "Tomato/Tomato", "Tomato/Tomato", "Tomato/Tomato")

## Differential analysis: DEseq2 method
# 24-nt sRNA data-set
sRNA_24_DESeq2 <- RNAanalysis(data = sRNA_24,
                             group = groups,
                             method = "DESeq2" )
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> final dispersion estimates
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> fitting model and testing
# 2122-nt sRNA data-set
sRNA_2122_DESeq2 <- RNAanalysis(data = sRNA_2122,
                               group = groups,
                               method = "DESeq2")
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> final dispersion estimates
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> fitting model and testing

## Differential analysis: edgeR method
sRNA_24_edgeR <- RNAanalysis(data = sRNA_24,
                            group = groups,
                            method = "edgeR" )
#> Design matrix not provided. Switch to the classic mode.

sRNA_2122_edgeR <- RNAanalysis(data = sRNA_2122 ,
                              group = groups,
                              method = "edgeR" )
#> Design matrix not provided. Switch to the classic mode.