Define the sRNA dicer consensus for each dicer-derived sRNA cluster

Using the data, the function uses the supplied dataframe and adds an additional column stating the consensus sRNA class for each dicer-derived cluster.

Usage

RNAdicercall(
  data,
  conditions = NULL,
  ties.method = NULL,
  tidy = FALSE,
  chimeric = FALSE,
  controls = NULL,
  genome.ID = NULL
)

Arguments

data

a data frame object containing sample data where rows represent sRNA dicer-derived clusters, and where columns represent sample data. See RNAimport() to load data, extract the required information for each sample and organise it as required.

conditions

character vector; containing names of sample replicates. Named replicates will be used to calculate dicercall consensus. Each string should represent a sample name present in the dataframe supplied to data argument.

ties.method

a character string specifying how ties are handled, "exclude" by default. Options include "random", or "exclude"; see ‘Details’

tidy

use of this argument will remove sRNA clusters with a unknown or unclassified consensus result. By default, the function will not tidy the data. When tidy=TRUE, unclassified clusters will be to remove from the output dataframe in an effort to remove excess background noise.

chimeric

logical; state whether system is chimeric: contains multiple genomes/genotypes.

controls

character; vector of control condition sample names.

genome.ID

character; chromosome identifier of foreign genome in chimeric system

Value

A data frame containing all existing columns in the input data object, plus, two additional columns of data:

The first column, DicerCounts contains the number of replicates which contributed to defining the dicer-derived sRNA class. This can be utilised within the RNAmobile function as a threshold parameter. While, utilising exclude ties methods, the DicerCounts value for a tie will be 0 as no consensus classification was concluded. While, if the majority of replicates contributed unclassified for a cluster, the DicerCounts value represents the number of replicates which which contributed to the unclassified consensus in DicerConsensus.

The second, labeled DicerConsensus states the consensus sRNA class between nucleotides in length or "N" if unclassified.

Details

The function calculates the consensus sRNA class. Changes in parameter settings will yield varying results.

When working with a chimeric system, for example interspecific grafting, mapping errors can easily be recognised and eliminated. Here, these can be eliminated by supplying some extra parameter information. State chimeric=TRUE and supply the chromosome identifier of the foreign genome (ie. not the tissue sample genotype, but the genotype from which any potential mobile molecules could be traveling from) to the genome.ID parameter & the control condition samples names to the controls parameter.

When ties.method = "random", as per default, ties are broken at random. In this case, the determination of a tie assumes that the entries are probabilities: there is a relative tolerance of 1e-5, relative to the largest (in magnitude, omitting infinity) entry in the row.

When ties.method = "exclude", ties between sRNA classification are ruled as unclassified ("N").

To remove excess data noise, tidy=TRUE can be used to removed unclassified ("N") sRNA clusters, resulting in a reduced dataset size.

Examples

 data("sRNA_data")

# define consensus sRNA classes.
conditions <- c("heterograft_1", "heterograft_2", "heterograft_3")

# Run function to define sRNA class for each cluster.
sRNA_data_consensus <- RNAdicercall(data = sRNA_data,
                                  conditions = conditions,
                                  tidy=TRUE)
#> Calculating consensus dicercall based on information from select replicates... 
#> 
#> The consensus dicercall will be excluded in the case of a tie... 
#> 
#> Removing small RNA clusters with no consensus dicercall...