Import and organise sRNAseq or mRNA data sets

Load and organise either sRNAseq or mRNAseq results into a single dataframe containing all experimental replicates specified where rows represent either a sRNA locus or gene, respectively.

Usage

RNAimport(
  input = c("sRNA", "mRNA"),
  directory,
  samples,
  report = TRUE,
  tidy = TRUE
)

Arguments

input

string; define type of Next-Generation Sequencing dataset originates from, either "sRNA" or "mRNA" are the only valid inputs.

directory

Path to directory containing of sample folders. NOTE: Following the suggested pre-processing steps, these can be found in second alignment folders.

samples

Vector of characters naming the sample names correlating to outputted folders located in the `directory` argument path.

report

Logical; prompts progress.

tidy

Logical; removes genes from analysis where there are zero counts across all replicates.

Value

A dataframe where rows represent sRNA clusters and columns represent replicate information. Replicate information includes Dicercall, Counts, RPM and MajorRNA sequence. Each replicate information is distinguishable as the replicate name is ajoined as a prefix to each column name.

Details

Supply the directory to the folder storing the various sample replicate folders produced by ShortStack analysis. Within each sample replicate folder, there is "Result.txt". This file contains many different columns, but for the analysis the columns of interest are "Locus", "DicerCall", "Reads", and "RPM" which will all be imported and loads into the new data frame. Note that "Reads" will be converted to "Counts" as output of the function. Locus contains the genomic locations of clusters, Dicercall contains the most likely size of the cluster, Reads contains the number of reads mapped to the cluster (recommended to use unique mapping, hence these will be the number of uniquely mapped reads, ie, not including multimapped reads). Lastly, RPM contains the Reads per Million score.

"RNAimport" takes the results from pre-processed sRNAseq data which has been mapped and undertaken cluster analysis through ShortStack, organises into a single data frame. Each row represent an sRNA dicer-derived cluster within the analysis and columns 1 - 5 supply information on the cluster including the locus, the separated coordinates (chromosome, start, end) and cluster name.

Further columns represent data imported for each samples including DicerCall, Counts and RPM. The DicerCall represents the size of most abundant small RNA size based on the parameter used in ShortStack. The Count column represents the number of aligned sRNA-seq reads that overlap the locus. The RPM represents the reads per million. For each replicate included in the analysis, these columns are labeled with the type and then then name of the sample, for example, for a sample called "Sample1", the information from this sample will be stored in columns DicerCall_Sample1, Count_Sample1 and RPM_Sample1.

Examples

if (FALSE) {

# import sRNAseq data
df_sRNA <- RNAimport(input = "sRNA",
                     directory = "./analysis/sRNA_mapping_results/",
                     samples = c("heterograft_1", "heterograft_2",
                     "heterograft_3","selfgraft_1" , "selfgraft_2" ,
                     "selfgraft_3"))

# import mRNAseq data
df_mRNA <- RNAimport(input = "mRNA",
                     directory = "./analysis/mRNA_counts/",
                     samples = c("heterograft_1", "heterograft_2",
                     "heterograft_3","selfgraft_1" , "selfgraft_2" ,
                     "selfgraft_3"))

}
# The output of this function can be explore in the data object sRNA_data
data("sRNA_data")
head(sRNA_data)
#>                   Locus chr     start       end   Cluster
#> 1 1:100000956-100001076   1 100000956 100001076 cluster_1
#> 2 1:100001486-100001554   1 100001486 100001554 cluster_2
#> 3 1:100001646-100001731   1 100001646 100001731 cluster_3
#> 4   1:10001034-10001073   1  10001034  10001073 cluster_4
#> 5 1:100011795-100011930   1 100011795 100011930 cluster_5
#> 6   1:10002000-10002065   1  10002000  10002065 cluster_6
#>   DicerCall_heterograft_1 Count_heterograft_1 RPM_heterograft_1
#> 1                      24                  19             1.908
#> 2                      24                   3             0.301
#> 3                       N                   3             0.301
#> 4                      24                   1             0.100
#> 5                      24                  11             1.105
#> 6                      24                   9             0.904
#>      MajorRNA_heterograft_1 DicerCall_heterograft_2 Count_heterograft_2
#> 1  AGUGACGGUCUCAUGUAACAUAGA                      24                  14
#> 2  AAAGAGUAGACAUAUAUUAAAUGG                      20                   5
#> 3 CUGUCUCAUGUGACACUCUUUCACU                      23                   6
#> 4  AAAUUGUCAUUUAGUCAGAGGAAU                       N                   0
#> 5  AUAGACAAAUCGCACUCGCUCUCG                      24                   2
#> 6  UUAGAUUUCUGGAGUAACAACUUU                       N                   2
#>   RPM_heterograft_2   MajorRNA_heterograft_2 DicerCall_heterograft_3
#> 1             1.803 AGUGACGGUCUCAUGUAACAUAGA                      24
#> 2             0.644     ACAUAUUUUUUUGGACGAAC                      24
#> 3             0.773  UCUGUCUCAUGUGACACUCUUUC                      24
#> 4             0.000                        N                      24
#> 5             0.258 AAGUGUCUACCCAUUGUGAUUCGC                      24
#> 6             0.258      UUAUUGACAAACUAUAACU                      22
#>   Count_heterograft_3 RPM_heterograft_3   MajorRNA_heterograft_3
#> 1                  30             2.953 AGUGACGGUCUCAUGUAACAUAGA
#> 2                   1             0.098 UAUGUCUACUCUUUAUUACUCUCC
#> 3                  11             1.083 AUCUUGUCAAGUGAAAGAGUGUCA
#> 4                  11             1.083 AUUCCUCUGACUAAAUGACAAUUU
#> 5                   5             0.492 AAUCACAAUGGGUAGACACUUAGA
#> 6                   3             0.295 UUAGAUUUCUGGAGUAACAACUUU
#>   DicerCall_selfgraft_1 Count_selfgraft_1 RPM_selfgraft_1
#> 1                    24                19           1.909
#> 2                    24                 3           0.301
#> 3                     N                 3           0.301
#> 4                    24                 1           0.100
#> 5                    24                11           1.105
#> 6                    24                 9           0.904
#>        MajorRNA_selfgraft_1 DicerCall_selfgraft_2 Count_selfgraft_2
#> 1  AGUGACGGUCUCAUGUAACAUAGA                    24                14
#> 2  AAAGAGUAGACAUAUAUUAAAUGG                    20                 5
#> 3 CUGUCUCAUGUGACACUCUUUCACU                    23                 6
#> 4  AAAUUGUCAUUUAGUCAGAGGAAU                     N                 0
#> 5  AAUGGGUAGACACUUAGAUACAUG                    24                 2
#> 6  UUAGAUUUCUGGAGUAACAACUUU                     N                 2
#>   RPM_selfgraft_2     MajorRNA_selfgraft_2 DicerCall_selfgraft_3
#> 1           1.804 AGUGACGGUCUCAUGUAACAUAGA                    24
#> 2           0.644     ACAUAUUUUUUUGGACGAAC                    24
#> 3           0.773  UCUGUCUCAUGUGACACUCUUUC                    24
#> 4           0.000                        N                    24
#> 5           0.258 AUAGACAAAUCGCACUCGCUCUCG                    24
#> 6           0.258  AGAUUUCUGGAGUAACAACUUUG                    22
#>   Count_selfgraft_3 RPM_selfgraft_3     MajorRNA_selfgraft_3
#> 1                30           2.954 AGUGACGGUCUCAUGUAACAUAGA
#> 2                 1           0.098 UAUGUCUACUCUUUAUUACUCUCC
#> 3                11           1.083 AUCUUGUCAAGUGAAAGAGUGUCA
#> 4                11           1.083 AUUCCUCUGACUAAAUGACAAUUU
#> 5                 5           0.492 AAUCACAAUGGGUAGACACUUAGA
#> 6                 3           0.295   UUGACAAACUAUAACUUAAAUU