Merge two FASTA reference genomes into one

Merges two reference genomes (.fa/.fasta). into one single reference with modified chromosome names to ensure distinguishability.

Usage

RNAmergeGenomes(
  genomeA,
  genomeB,
  out_dir,
  abbreviationGenomeA = "A",
  abbreviationGenomeB = "B",
  replace_chr_names = TRUE
)

Arguments

genomeA

the path base::connection() to a genome reference file in FASTA format (.fa/.fasta). File may be supplied as compressed with a .gzip extension.

genomeB

the path base::connection() to a genome reference file in FASTA format (.fa/.fasta). File may be supplied as compressed with a .gzip extension.

out_dir

either a character string or a base::connections() open for writing. Place path to output directory in "", including file output name.

abbreviationGenomeA

a string placed in "", to replace chromosome names within genomeA.Default set as "A".

abbreviationGenomeB

a string placed in "", to replace chromosome names within genomeB.Default set as "B".

replace_chr_names

a logical value indicating whether the chromosome names of the supplied annotation files are to be altered or not. As default, chromosome names are altered.

Value

A FASTA format file containing two genome assemblies with edited chromosome names (prefixes, and removal of periods)

Details

The functions primary goal is to merge two FASTA files, however, when merging genomic files it is critical that the two genomes are distinguishable by the chromosome names. As a default setting, the function extracts the chromosome names for a given FASTA file and alters the name while retaining the identifying number.

The function requires the input of two reference genomes, where one represents Genome-A and the other represents Genome-B. As default, the function will rename the chromosome names in Genome-A to "A". For example, A0, A1, A2 etc. To set a custom chromosome name for Genome-A alter the argument abbreviationGenomeA. While, for Genome-B as default the chromosome names will be named "B", for example, B0, B1, B2 etc. To set a custom chromosome name for Genome-A alter the argument abbreviationGenomeB. The function can do so by draw the chromosome number within the given GFF file, remove all prior character or numerical values, and replace it with the supplied string.

IMPORTANT: The genome reference and annotation of a species must have chromosomes with matching names. It is critical that if you used the RNAmergeAnnotations() function to to create a merged genome annotation,that you treat the input references in the same way.

Examples

if (FALSE) {
merged_ref <- RNAmergeGenomes(genomeA = "./workplace/reference/ref1.fa",
            genomeB = "./workplace/reference/ref2.fa",
            out_dir = "./workplace/reference/merge/merged_ref.fa")

## or, to set specific changes to chromosome names. annotationA represents
## the Solanum lycopersicum and the chromosomes will be abbreviated to `SL`,
## and annotationB represents Solanum melongena and the chromosomes will be
## abbreviated to `SM`.

merged_ref2 <- RNAmergeGenomes(genomeA = "./workplace/reference/ref1.fa",
            genomeB = "./workplace/reference/ref2.fa",
            out_dir = "./workplace/reference/merge/merged_ref.fa",
            abbreviationGenomeA = "SL",
            abbreviationGenomeB = "SM")

}