Identify potential mobile sRNA & mRNA molecules

A function to identify potential mobile sRNA or mRNA moleculea detected in a tissue which originated from a different genotype, usually related to a chimeric system.

Usage

RNAmobile(
  data,
  controls,
  genome.ID,
  task = NULL,
  statistical = FALSE,
  padj = 0.05,
  threshold = NULL,
  input = NULL,
  p.value = NULL
)

Arguments

data

Numeric data frame

controls

Character vector; containing names of control samples.

genome.ID

a character string related to the chromosomes in a particular genome. A distinguishing feature of the genome of interest or non-interest in the chromosome name (chr column).

task

an option to keep or remove the chromosomes containing the identifying string. To keep the chromosomes with the ID, set task=keep. To remove, set task="remove". As default, task is set to keep.

statistical

If TRUE, will undertake statistical filtering based on the a p-value or adjusted p-value threshold stated by padj & p.value.Default set at FALSE. Requires presence of columns containing statistical data. In order to filter by the adjusted p-value, a column named padjusted must be present. Similarly, to filter by the p-value, a column named pvalue must be present. See RNAanalysis() to calculate statistical values.

padj

A user defined numeric value to represent the adjusted p-value threshold to define statistic significance. Defaults set at 0.05.Only mobile molecules with adjusted p-values equal or lower than specified are returned.

threshold

numeric; set a threshold level. For sRNAseq, this represents filtering by the minimum number of replicates that defined the dicer-consensus which is stored in the DicerCounts column. For mRNAseq, this represents the minimum number of replicates which contained reads at the gene locus, this information is storeds in the SampleCounts column.

input

character; must be either "sRNA" or "mRNA" to represent the type of data, required when setting theshold.

p.value

A user defined numeric value to represent the p-value threshold to define statistic significance. There is no default value, set this instead of using an adjusted p-value to filter molecules. Only mobile molecules with p-values equal or lower than specified are returned.

Value

A data-frame containing candidate mobile sRNAs, which could be further filtered based on statistical significance and the ability to by-pass the threshold which determines the number of replicates that defined the dicer-consensus.

Details

For sRNAseq: The function identifies candidate mobile sRNAs, by selecting sRNA clusters mapped to the foreign genome. It does so by either keeping or removing sRNA mapped to a given genome. To do so, it requires a common pre-fix across chromosomes of the same genome. See RNAmergeGenomes() for more information.

A greater confidence in the mobile sRNA candidates can be achieved by setting a threshold that considers the number of replicates which contributed to defining the consensus dicer-call (ie. consensus sRNA classification). This parameter filters based on the DicerCounts column introduced by the RNAdicercall() function.

For mRNAseq: The function identifies candidate mobile mRNAs, by selecting mRNA molecules mapped to the foreign genome. It does so by either keeping or removing mRNA mapped to a given genome. To do so, it requires a common pre-fix across chromosomes of the same genome. See RNAmergeGenomes() and RNAmergeAnnotations() for more information and assistance.

A greater confidence in the mobile mRNA candidates can be achieved by setting a threshold that considers the number of replicates which contained reads for the mRNA molecule. This parameter filters based on the SampleCounts column introduced by the RNAimport() function.

Statistical Analysis The function also allows for statistical analysis based on the results collect from differential analysis of the total dataset using the function RNAanalysis(). When statistical=TRUE, the feature is enabled and selected mobile molecules that meet a specific p-value or adjusted p-values.

Examples

data("sRNA_data_consensus")

# vector of control names
controls <- c("selfgraft_1", "selfgraft_2" , "selfgraft_3")

# Locate potentially mobile sRNA clusters associated to tomato, no
# statistical analysis
mobile_df1 <- RNAmobile(data =  sRNA_data_consensus,
                    controls = controls,
                    genome.ID = "SL40",
                    task = "keep",
                    statistical = FALSE)