Extract sRNA cluster sequences

Extract sRNA cluster sequences; identifies whether the most abundant sRNA for a sRNA cluster is consistent across a defined number of replicates, and if so extracts the sRNA nucleotide sequence.

The function also calculates the RNA & DNA complementary sequences, as well as stating the length of the sequence.

The results can be stored in the the working dataframe supplied to the function or stored as a new dataframe containing only the results from the function. This includes:

Match: whether the sRNA sequence is consistent across replicates

Sequence: character; sequence of the most abundant sRNA within a cluster

Complementary_RNA: character; complementary RNA nucleotide sequence

Complementary_DNA: character; complementary DNA nucleotide sequence

Width: numeric; length of nucleotide sequence

Usage

RNAsequences(
  data,
  original = FALSE,
  match.threshold = 1,
  duplicates = "random",
  tidy = FALSE
)

Arguments

data

dataframe; containing columns starting with Cluster & MajorRNA

original

logical; output results to working date as additional columns (original=TRUE), or output as new dataframe (original=FALSE). by default, FALSE

match.threshold

numeric; the minimum number of replicates required to share the sRNA sequence to count as a match

duplicates

character string to define how to deal with a tie, "random" as default. Options include "random" and "exclude"

tidy

use of this argument will remove sRNA clusters with a unknown or unclassified consensus major sRNA sequence result. By default, tidy=FALSE, but while tidy=TRUE sRNA clusters with undetermined consensus will be removed.

Value

The results can be stored in the the working dataframe supplied to the function or stored as a new dataframe containing only the results from the function. The results includes:

Match: whether the sRNA sequence is consistent across replicates

Sequence: character; sequence of the most abundant sRNA within a cluster

Complementary_RNA: character; complementary RNA nucleotide sequence

Complementary_DNA: character; complementary DNA nucleotide sequence

Width: numeric; length of nucleotide sequence

Details

As default, the function checks whether each sample in the dataset shares the same major sRNA sequence for a given sRNA cluster. If at least two replicates share the same sRNA sequence, the sequence is pulled and the complementary DNA and RNA sequences are calculated. Using the match.threshold parameter, we can alter the minimum number of replicates required to share the sRNA sequence to count as a match. For example, if set as match.threshold=3, at least 3 replicates must contain the same sequence. As a general rule, if only one replicate has determined a sRNA sequence it is noted that there is no match, but the sequence is pulled and the complementary sequences calculated.

The match column can either return "Yes", "No" or "Duplicate". If a match between replicates is found, "Yes" is supplied, if not, "No". While if there is a tie between sequences "Duplicate" is supplied. For examples, if an equal number of replicates have sequence "x" and sequence "y".

In the situation where duplicates are identified, as default, at random a consensus sRNA sequence is selected. This parameter can be changed to "exclude", and under this parameter no consensus sequence is pulled.

Examples

data('sRNA_data_mobile')

mobile_sequences <- RNAsequences(sRNA_data_mobile)
#> The minimum number of matches required to form a consensus sRNA sequence is...  1 
#> The consensus sRNA sequences will be choose at random in the case of a tie 
#>