Differential Expression (DE) Analysis of sRNA dicer-derived clusters using DESeq2 or edgeR

This function allows you to compute the differential expression of sRNA dicer-derived clusters. The function allows the choice between analysis with DESeq2 or edgeR.

Usage

RNAanalysis(data, group, method = c("edgeR", "DESeq2"), dispersionValue = NULL)

Arguments

data

numeric data frame produced by RNAimport() and/or RNAsubset().

group

Vector of the condition (ie. treatment or control) for each sample. Must be stated in the same order as the samples in the data file from left to right.

method

The method to undertaken differential analysis, choose from methods of either DESeq2::DESeq or edgeR::edgeR. Must be stated as either "DESeq2" or "edgeR" in the function.

dispersionValue

numeric value; manual setting of dispersion value which is recommended for analysis in experiments without biological replicates when utilising the edgeR::edgeR method.

Value

Undertakes differential analysis, based on a specified method, and adds the results to the supplied dataframe. This includes:

Count mean

Log fold change

p-value

Adjusted p-value

Log counts per million (CPM/RPM)

Details

The analysis allows the users to choose the method which best suits their data. Notably, DESeq2 cannot compute the analysis when there only one replicate per condition, but, edgeR can. Simply set a suitable dispersion value, based on similar data, to use this feature. The dispersion value is other wise known as the common Biological squared coefficient of variation. A typical dispersion value is 0.4 for human data sets, 0.1 for data on genetically identical model organisms or 0.01 for technical replicate. See the User’s Guide for the ‘EdgeR’ package for more details, edgeR::edgeR.

Examples

data("sRNA_data_consensus")
# sample conditions.
groups <- c("Heterograft", "Heterograft", "Heterograft",
          "Selfgraft", "Selfgraft", "Selfgraft")

## Differential analysis: DEseq2 method
sRNA_DESeq2 <- RNAanalysis(data = sRNA_data_consensus,
                             group = groups,
                             method = "DESeq2" )
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing

## Differential analysis: edgeR method
sRNA_edgeR <- RNAanalysis(data = sRNA_data_consensus,
                            group = groups,
                            method = "edgeR" )
#> Design matrix not provided. Switch to the classic mode.