Merge two GFF genome annotation files

Merges two genomes annotation files (GFF format) into one single GFF file saved to desired directory. Adds prefix to each chromosome name in the genome annotations to ensure each assembly is distinguishable within the merged file, and saved the individually altered references to local directory.

Usage

RNAmergeAnnotations(
  annotationA,
  annotationB,
  out_dir,
  abbreviationAnnoA = "A",
  abbreviationAnnoB = "B"
)

Arguments

annotationA

GRanges object; a genome annotation assembly file in GFF format.

annotationB

GRanges object; a genome annotation assembly file in GFF format.

out_dir

either a character string or a base::connections() open for writing. Place path to output directory in "", including file output name with extension of either .gff or .gff3.

abbreviationAnnoA

a string placed in "", to replace chromosome names within annotationA.Default set as "A".

abbreviationAnnoB

a string placed in "", to replace chromosome names within annotationB.Default set as "B".

Value

The function output the individually altered annotation files, plus, a merged annotation file to the global environment and the given directory. The altered annotation files contain either default or custom alterations to the chromosome names and are written to the same directory as the merged file, with the given names addition of annotationA_altered.gff and annotationB_altered.gff, respectively.

Details

The functions primary goal is to merge two GFF files, however, when merging genomic files it is critical that the two genomes are distinguishable by the chromosome names. As a default setting, the function extracts the chromosome names for a given GFF file and adds an ID before each chromosome name and removes any periods.

As default, the function will rename the chromosome names in annotationA to "A". For example, A0, A1, A2 etc. To set a custom chromosome name for annotationA alter the argument abbreviationAnnoA. While, for annotationB as default the chromosome names will be named "B", for example, B0, B1, B2 etc. To set a custom chromosome name for annotationB alter the argument abbreviationAnnoB.

The merged genome is saved to the specified output directory, and requires the user to set the name with a gff extension. The user must load the annotations into the R Global Environment, hence, the alterations made by this function will alter the variable. The altered annotations will also be saved to the same directory as supplied for the output file, with a fixed output name of annotationA_altered.gff and annotationB_altered.gff, respectively.

IMPORTANT: The genome reference and annotation of a species must have chromosomes with matching names. It is critical that if you used the RNAmergeGenomes() function to to create a merged reference genome,that you treat the input annotations in the same way.

Examples

# import GFF files into R

# Read reference genomes
url_remote <- "https://github.com/KJeynesCupper/assemblies/raw/main/"

anno1_url <- paste0(url_remote,"chr12_Eggplant_V4.1_function_IPR_final.gff")

anno2_url <- paste0(url_remote, "chr2_ITAG4.0_gene_models.gff")

 # Read in GFF files
anno1 <- rtracklayer::import(anno1_url)
anno2 <- rtracklayer::import(anno2_url)

# edit and merge
merged_anno <- RNAmergeAnnotations(annotationA = anno1,
                                 annotationB = anno2,
             out_dir = "./merged_annotation.gff3")
#> Adding abbreviations to chomosome names ... 
#> New annotation file created: ./annotationA_altered.gff
#> New annotation file created: ./annotationB_altered.gff
#> Merging altered annotation files ...

## Set specific pre-fixes:
###  annotationA represents Solanum melongena, so abbreviated to `SM`
###  annotationB represents Solanum lycopersicum, so abbreviated to `SL`

merged_anno_2 <- RNAmergeAnnotations(annotationA = anno1,annotationB = anno2,
              out_dir = "./merged_annotation.gff3",
              abbreviationAnnoA = "SM",abbreviationAnnoB = "SL")
#> Adding abbreviations to chomosome names ... 
#> New annotation file created: ./annotationA_altered.gff
#> New annotation file created: ./annotationB_altered.gff
#> Merging altered annotation files ...